apps/api/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.0

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6

# Google integration
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1
google-api-python-client==2.111.0

# Task queue
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1